[Refactor] (test) Refactor the testing items about registrations to new testing module 'register'.
